+++ R/copy.R
@@ -34,8 +34,8 @@ copyVariable <- function (x, deep=FALSE, ...) {
     body <- modifyList(oldbody, newbody)
     body$id <- NULL
     if (deep) {
-        if (body$type %in% c("categorical_array", "multiple_response")) {
-            halt("Deep copying of array variables is not implemented.")
+        if (body$type == "multiple_response") {
+            halt("Deep copying of multiple-response variables is not implemented.")
         }
         body$values <- as.vector(x, mode="id")
         if (body$type %in% c("categorical", "categorical_array", "multiple_response")) {
@@ -44,7 +44,9 @@ copyVariable <- function (x, deep=FALSE, ...) {
                 body$subvariables_catalog <- NULL
                 body$subvariables <- lapply(names(subvariables(x)),
                     function (n) list(name=n))
-                ## Format the values?
+                ## Turn values into a matrix, rather than data.frame
+                body$values <- matrix(do.call("c", body$values),
+                    nrow=nrow(body$values))
             }
         } else if (body$type == "datetime") {
             body$resolution <- entity(x)@body$resolution